@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import { CookieService } from 'ngx-cookie-service';
 
 @Component({
   selector: 'app-home',
@@ -8,13 +9,14 @@ import { Router } from '@angular/router';
 })
 export class HomeComponent {
 
-  constructor(private router: Router) {
+  constructor(private router: Router, private cookieService: CookieService) {
 
   }
 
   ngOnInit() {
-    console.log(localStorage.getItem("jwt"))
-
+    if (this.cookieService.get("jwt") === null || this.cookieService.get("jwt") === "") {
+      this.router.navigate(['/login']);
+    }
   }
 
 }
